RF: Cleanly separate metadata extraction, query, and aggregation code
